Switch to git for OpenMP fixes.
This switch effectively moves us to what is, in essence, 3.0.0. This is a major release with database changes, is not backwards compatible, and includes a significant amount of changes. We'll switch to 3.0.0 when it's properly released, for now view the release notes [here](https://github.com/darktable-org/darktable/blob/master/RELEASE_NOTES.md).